refactor(phase4): extract storage_parsing from eon_parser.rs
Move `build_static_mounts`, `parse_storage_document`, mount-path
validators, storage-backend helpers, and `validate_storage_upload_routes`
(~642 lines) into `eon_parser/storage_parsing.rs`.  Prune corresponding
stale imports from `eon_parser.rs`.
All 233 passing tests remain green; 19 pre-existing failures unchanged.
